job/requirements.txt
# core
openai>=1.40.0
dashscope>=1.20.0

# db
pymongo[srv]>=4.6.0

# r2 (s3-compatible)
boto3>=1.28.0

# http
requests>=2.31.0
httpx>=0.27.0

# scraping / parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.0.0

# nlp
ltp==4.2.14
numpy<2
pypinyin>=0.50.0

# audio
pydub>=0.25.1

# data (only if still used)
pandas>=2.2.0

# local env loading (optional in cloud, useful locally)
python-dotenv>=1.0.0